# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Samoey Copilot API
  description: Comprehensive API documentation for Samoey Copilot project
  version: 1.0.0
  contact:
    name: Samoey Copilot Team
    email: support@samoey-copilot.com
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.samoey-copilot.com/v1
    description: Production server
paths:
...
  /users/:
    get:
      tags:
        - Users
      summary: Get all users
      description: Retrieve a list of all users (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user
      description: Get information about the current authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        full_name:
          type: string
        is_active:
          type: boolean
          default: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...